separate shoe and shoes reducer
