+++ src/reducers/rootReducer.js
@@ -1,9 +1,11 @@
 import { combineReducers } from 'redux';
-import shoeReducer from './shoesReducer';
+import shoesReducer from './shoesReducer';
+import shoeReducer from './shoeReducer';
 import authReducer from './authReducer';
 import favouriteReducer from './favouritesReducer';
 
 const rootReducer = combineReducers({
+  shoes: shoesReducer,
   shoe: shoeReducer,
   auth: authReducer,
   fav: favouriteReducer,
